requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.1.6
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-auth==2.23.2
google-auth-oauthlib==1.1.0
gspread==5.11.3
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-decouple==3.7
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
